+++ content/blogs/2026-04-14_Why-Does-Using-AI-To-Write-Feel-Like-Cheating.mdx
@@ -36,13 +36,13 @@ Still, I don't think it is that simple. The discomfort seems to track with how h
 
 Nobody really blinks at AI helping with code. A lot of people are comfortable with AI assisting scientific research, protein folding, drug discovery, and data analysis. Those feel like tool problems. But AI-generated art gets a more visceral reaction. AI music can feel hollow even when people cannot clearly explain why. AI writing sits somewhere in the middle, and the reaction shifts depending on the kind of writing.
 
-There seems to be a line somewhere, and it tends to sit near whatever we treat as self-expression. The closer the output gets to something that feels personal, the more the assistance starts to feel intrusive.
+There seems to be a line somewhere, and I think it tends to sit near whatever we treat as self-expression. The closer the output gets to something that feels personal, the more the assistance starts to feel intrusive.
 
-A calculator does not threaten anyone's identity. Neither did a thesaurus, a spell checker, or a style guide. But this new tool that writes your prose does for many.
+A calculator doesn't threaten anyone's identity. Neither did a thesaurus, a spell checker, or a style guide. But this new tool that writes your prose does for many.
 
 ## I Don't Know Where the Line Is
 
-I can feel that tension in myself, and I do not have a clean answer for it.
+I can feel that tension in myself, and I don't have a clean answer for it.
 
 The argument I made above, that the ideas are mine and the AI only helps with execution, feels mostly true for technical writing. I am explaining things I have thought and built and patterns I have seen firsthand. In that context, the LLM feels more like a drafting tool than a ghostwriter.
 
@@ -58,7 +58,7 @@ What I do know is simpler than all of that: For years, I've had ideas sitting in
 
 For me, right now, that is still a difficult choice, but one I am going to make differently now.
 
-The ideas in my recent posts are mine. The examples come from my own experience. The opinions are things I have tested in production. An LLM helped me get past the block and organize my thinking faster. I am comfortable with that.
+The ideas in my posts are going to be mine. The examples come from my own experience. The opinions are things I have tested in production. An LLM helped me get past the block and organize my thinking faster. I am comfortable with that.
 
 Whether that comfort extends to every kind of writing, I am less certain. Anyone who claims to know exactly where AI stops being a tool and starts becoming something else is probably drawing the line in a place that suits them.
